remove old code, update localforage versions
