--- WebScience/dependencies/localForage-startsWith/lib/implementations/idbKeyRange.js
@@ -1,15 +1,0 @@
-function getIDBKeyRange() {
-    /* global IDBKeyRange, webkitIDBKeyRange, mozIDBKeyRange */
-    if (typeof IDBKeyRange !== 'undefined') {
-        return IDBKeyRange;
-    }
-    if (typeof webkitIDBKeyRange !== 'undefined') {
-        return webkitIDBKeyRange;
-    }
-    if (typeof mozIDBKeyRange !== 'undefined') {
-        return mozIDBKeyRange;
-    }
-}
-
-var idbKeyRange = getIDBKeyRange();
-export default idbKeyRange;
