--- settings/Options.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-
-<html>
-    <head>
-        <meta charset="utf-8">
-        <link rel='stylesheet' href='Options.css'/>
-    </head>
-
-    <body>
-        <!-- TODO: add information about the larger Pioneer project/links to more info-->
-        <!-- TODO: once we've settled on a differential privacy setup, add info about how
-        data collection is anonymized/sent/etc -->
-        <p>This study collects anonymized information about your consumption of news online, as well
-            as links you see to news stories.</p>
-
-        <p>You can opt-out by disabling the study below, which will uninstall the extension that
-            runs the study.</p>
-
-        <p>If you'd just like to browse temporarily without your browsing data being used for the study,
-            you can use a Private Window.</p>
-
-        <form>
-            <b>Study enabled</b>
-            <label class="switch">
-                <input type="checkbox" id="collection">
-                <span class="slider round"></span>
-            </label>
-            <button type="submit">Save</button>
-        </form>
-
-        <script type="module" src="Options.js"></script>
-
-    </body>
-
-</html>
