--- settings/Options.js
@@ -1,33 +1,0 @@
-async function saveOptions(e) {
-  e.preventDefault();
-  /* When the user hits the save button, send a message
-   *  to the background page with the new value.
-   */
-  var consent = document.querySelector("#collection").checked;
-
-  browser.runtime.sendMessage({
-    type: "WebScience.Options.saveStudySpecificConsent",
-    studySpecificConsent: consent
-  });
-}
-
-async function restoreOptions() {
-
-  function setCurrentChoice(result) {
-    document.querySelector("#collection").checked = result;
-  }
-
-  /* Send a message to the background page to ask whether
-   *  the study is currently enabled, and then set the slider
-   *  to reflect that -- the response from the listener is
-   *  the answer.
-   * Note: expect the listener in Consent.js to be the
-   *  only one to provide a response.
-   */
-  browser.runtime.sendMessage({
-    type: "WebScience.Options.checkStudySpecificConsent"
-  }).then(setCurrentChoice);
-}
-
-document.addEventListener("DOMContentLoaded", restoreOptions);
-document.querySelector("form").addEventListener("submit", saveOptions);
