+++ study/WebScience.Utilities.js
@@ -12,9 +12,6 @@ export { Storage }
 import * as Messaging from "../WebScience/Utilities/Messaging.js"
 export { Messaging }
 
-import * as Lifecycle from "../WebScience/Utilities/Lifecycle.js"
-export { Lifecycle }
-
 import * as Idle from "../WebScience/Utilities/Idle.js"
 export { Idle }
